Remove FreeRTOS code from the source tree
No need to keep all the FreeRTOS sources in the projects source tree as this project makes
use of unmodified version of FreeRTOS. The only thing that needs to be kept locally is
include/FreeRTOSConfig.h.

Note 1.
Before building this project one have to modify Makefile and to set FREERTOS_SOURCE_DIR properly.

Note 2.
So far tested with FreeRTOS v7.3.0